Feedback 3
- Separated test classes
- Typos
- Reverted Error Message
