+++ src/NuGet.Clients/NuGet.PackageManagement.UI/LoadingStatusIndicator.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.ComponentModel;
@@ -11,6 +11,7 @@ namespace NuGet.PackageManagement.UI
         public event PropertyChangedEventHandler PropertyChanged;
 
         private LoadingStatus _status = LoadingStatus.Unknown;
+        private string _errorMessage;
         private string _loadingMessage;
 
         public LoadingStatus Status
@@ -39,6 +40,22 @@ namespace NuGet.PackageManagement.UI
             }
         }
 
+        public string ErrorMessage
+        {
+            get
+            {
+                return _errorMessage;
+            }
+            set
+            {
+                if (_errorMessage != value)
+                {
+                    _errorMessage = value;
+                    OnPropertyChanged(nameof(ErrorMessage));
+                }
+            }
+        }
+
         protected void OnPropertyChanged(string propertyName)
         {
             if (PropertyChanged != null)
@@ -57,6 +74,12 @@ namespace NuGet.PackageManagement.UI
         {
             Status = LoadingStatus.Unknown;
             LoadingMessage = loadingMessage;
+        }
+
+        public void SetError(string message)
+        {
+            Status = LoadingStatus.ErrorOccurred;
+            ErrorMessage = message;
         }
     }
 }
