+++ test/NuGet.Clients.Tests/NuGet.PackageManagement.UI.Test/Xamls/InfiniteScrollListTests.cs
@@ -8,12 +8,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.Threading;
 using Moq;
-using NuGet.Common;
 using NuGet.PackageManagement.VisualStudio;
-using NuGet.Packaging;
-using NuGet.Packaging.Core;
 using NuGet.ProjectManagement;
-using NuGet.Protocol;
 using NuGet.Protocol.Core.Types;
 using NuGet.VisualStudio;
 using Xunit;
@@ -354,122 +350,6 @@ namespace NuGet.PackageManagement.UI.Test
             };
 
             return allData;
-        }
-
-        /// <summary>
-        /// Test implementation of <see cref="IPackageSearchMetadata"/>. Not for produciton use
-        /// </summary>
-        class TestPackageSearchMetadata : IPackageSearchMetadata
-        {
-            public string Authors { get; set; }
-
-            public IEnumerable<PackageDependencyGroup> DependencySets { get; set; }
-
-            public string Description { get; set; }
-
-            public long? DownloadCount { get; set; }
-
-            public Uri IconUrl { get; set; }
-
-            public PackageIdentity Identity { get; set; }
-
-            public Uri LicenseUrl { get; set; }
-
-            public Uri ProjectUrl { get; set; }
-
-            public Uri ReportAbuseUrl { get; set; }
-
-            public Uri PackageDetailsUrl { get; set; }
-
-            public DateTimeOffset? Published { get; set; }
-
-            public string Owners { get; set; }
-
-            public bool RequireLicenseAcceptance { get; set; }
-
-            public string Summary { get; set; }
-
-            public string Tags { get; set; }
-
-            public string Title { get; set; }
-
-            public bool IsListed { get; set; }
-
-            public bool PrefixReserved { get; set; }
-
-            public LicenseMetadata LicenseMetadata { get; set; }
-
-            public Task<PackageDeprecationMetadata> GetDeprecationMetadataAsync()
-            {
-                throw new NotImplementedException();
-            }
-
-            public Task<IEnumerable<VersionInfo>> GetVersionsAsync()
-            {
-                throw new NotImplementedException();
-            }
-        }
-
-        /// <summary>
-        /// NuGet UI Logger to print messages to Xunit test output
-        /// </summary>
-        class TestNuGetUILogger : INuGetUILogger
-        {
-            private readonly ITestOutputHelper _out;
-
-            public TestNuGetUILogger(ITestOutputHelper outputHelper)
-            {
-                _out = outputHelper;
-            }
-
-            public void End()
-            {
-                throw new NotImplementedException();
-            }
-
-            public void Log(MessageLevel level, string message, params object[] args)
-            {
-                _out.WriteLine($"[{level}] {string.Format(message, args)}");
-            }
-
-            public void Log(ILogMessage message)
-            {
-                Log(FromLogLevel(message.Level), message.Message);
-            }
-
-            public void ReportError(string message)
-            {
-                Log(MessageLevel.Error, message);
-            }
-
-            public void ReportError(ILogMessage message)
-            {
-                Log(MessageLevel.Error, message.Message);
-            }
-
-            public void Start()
-            {
-                throw new NotImplementedException();
-            }
-
-            private MessageLevel FromLogLevel(LogLevel logLevel)
-            {
-                switch (logLevel)
-                {
-                    case LogLevel.Error:
-                        return MessageLevel.Error;
-                    case LogLevel.Warning:
-                        return MessageLevel.Warning;
-                    case LogLevel.Verbose:
-                    case LogLevel.Debug:
-                        return MessageLevel.Debug;
-                    case LogLevel.Minimal:
-                    case LogLevel.Information:
-                        return MessageLevel.Info;
-                    default:
-                        return MessageLevel.Warning;
-                }
-            }
         }
     }
 }
